test/requirements.txt
flask
redis
kafka-python
faker
metrobus
